wishlist 기능 구현

+++ LTM/src/main/java/com/ltm/web/controller/PlayListController.java
@@ -7,6 +7,7 @@ import javax.persistence.EntityNotFoundException;
 import javax.validation.Valid;
 
 import org.springframework.data.domain.Page;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -17,11 +18,13 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.ltm.web.Dto.PlayListFormDto;
 import com.ltm.web.Service.MemberService;
 import com.ltm.web.Service.PlSongService;
 import com.ltm.web.Service.PlayListService;
+import com.ltm.web.Service.WishListService;
 import com.ltm.web.api.SongImageApi;
 import com.ltm.web.entity.Member;
 import com.ltm.web.entity.playlist.PlSong;
@@ -42,6 +45,7 @@ public class PlayListController {
 	private final MemberService memberService;
 	private final PlSongRepository plSongRepository;
 	private final SongImageApi songImageApi;
+	private final WishListService wishListService;
 
 	// 플레이리스트 만들기
 	@PreAuthorize("isAuthenticated()")
@@ -120,13 +124,8 @@ public class PlayListController {
 		// 쓰지않고 서비스를 통해서 사용하도록 작성
 		// model.addAttribute("playList2", playList);
 
-		System.out.println("***************************");
-		System.out.println(page);
-		System.out.println(kw);
-		System.out.println("***************************");
-
 		if ("".compareTo(kw) == 0) {
-			Page<PlayList> paging = this.playListService.getlist(page);
+			Page<PlayList> paging = this.playListService.getlist(page);	
 			model.addAttribute("paging", paging);
 			model.addAttribute("kw", kw);
 			return "playlist/Pl_main";
@@ -140,11 +139,26 @@ public class PlayListController {
 	}
 
 	//플레이리스트 상세 페이지
+	@PreAuthorize("isAuthenticated()")
 	@GetMapping("/{id}/song")
-	public String pldetail(Model model, @PathVariable("id") Long plId) {
+	public String pldetail(Model model, @PathVariable("id") Long plId, Principal principal) {
+		
+		//로그인한 회원 조회
+		Member member = this.memberService.getMember(principal.getName());
+		//그 회원의 위시리스트 조회
+		List<PlayList> findWl = wishListService.findWl(member.getIdNum());
+		
 		PlayList playlist = this.playListService.findOne(plId);
-		model.addAttribute("playList22", playlist);
+		
+		for (int i = 0; i < findWl.size(); i++) {
+			if(findWl.get(i).getId() != playlist.getId()) {
+				model.addAttribute("findWl",findWl.get(i).getId());
+			}
+		}
 
+		model.addAttribute("playList22", playlist);
+		
+		
 		List<PlSong> songs = plSongService.findPlSongs(plId);// 리스트로 담는 것도 생각해보기, plsong 연결필요
 		model.addAttribute("song22", songs);
 		return "playlist/Pl_detail";
@@ -200,11 +214,17 @@ public class PlayListController {
 	//플레이리스트에서 노래삭제
 	@PreAuthorize("isAuthenticated()")
 	@GetMapping("/delete/{plId}/{id}")
-	public String plSongDelete(@PathVariable("plId") Integer plId, @PathVariable("id") Integer id) {
+	public String plSongDelete(@PathVariable("plId") Integer plId, @PathVariable("id") Integer id, Principal principal) {
 		System.out.println("4444444444444");
+		
+		
 		
 //		PlayList playList = this.playListService.getPl(Long.valueOf(plId));
 		PlSong plSong = this.playListService.getPlsong(Long.valueOf(id));
+		
+		if(!plSong.getPlayList().getMember().getUsername().equals(principal.getName())) {
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제권한이 없습니다.");
+		}
 		
 		this.playListService.deletePlsong(plSong);
 		return "redirect:/main";
